Relocate IFW filter summary logic to feature module
- Move `IfwFilter.toSummary()` from `core:ifw-api` to `feature:globalifwrule`
- Rename the imported extension to `toIntentFilterSummary` in `GlobalIfwRuleMapper` to avoid name collisions
- Update internal visibility for the relocated summary utility functions
